+++ src/views/accounts/Account.jsx
@@ -1,45 +1,21 @@
-import { Hide, Icon, Link, Text } from '@chakra-ui/react'
+import { Text } from '@chakra-ui/react'
 import { useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 
 import { ProfilesAPI } from '@api'
 import {
-  BlockIcon,
   ChangeEmailCard,
   Container,
   DeleteAccountCard,
   InactivateAccountCard,
   Loading,
   LoadingBackdrop,
-  MessageIcon,
   ProfileHeader,
-  TrashIcon,
 } from '@components'
 import { StorageService } from '@services'
 
 import { useSelector } from 'react-redux'
 import style from './Account.module.css'
-
-const linksOptions = [
-  {
-    id: 'email',
-    label: 'Alterar e-mail',
-  },
-  {
-    id: 'inactive',
-    label: 'Inativar conta',
-  },
-  {
-    id: 'delete',
-    label: 'Excluir conta',
-  },
-]
-
-const icons = {
-  email: MessageIcon,
-  inactive: BlockIcon,
-  delete: TrashIcon,
-}
 
 export default function AccountView() {
   const navigate = useNavigate()
@@ -47,30 +23,6 @@ export default function AccountView() {
 
   const [fetching, setFetching] = useState(true)
   const [profile, setProfile] = useState(null)
-
-  const renderLinks = () => {
-    return (
-      <Container gap={22} column>
-        {linksOptions.map((linkItem) => (
-          <Link
-            key={linkItem.id}
-            href={'#' + linkItem.id}
-            color="gray.300"
-            fontWeight="regular"
-            fontSize={18}
-            lineHeight={1}
-            display="flex"
-            alignItems="center"
-            gap="10px"
-            className={style.link}
-          >
-            <Icon as={icons[`${linkItem.id}`]} fill="gray.300" />
-            <Text>{linkItem.label}</Text>
-          </Link>
-        ))}
-      </Container>
-    )
-  }
 
   useEffect(() => {
     const fetch = async () => {
@@ -100,27 +52,25 @@ export default function AccountView() {
         <ProfileHeader profile={profile} isUserLogged={true} />
       </Container>
 
-      <Container className={style.content}>
-        <Hide above="2xl">
-          <Container column style={{ maxWidth: '350px' }}>
-            <Container fitContent style={{ marginBottom: '40px' }}>
-              <Text
-                color="white"
-                fontWeight="bold"
-                fontSize={20}
-                textTransform="uppercase"
-              >
-                Gerenciamento de conta
-              </Text>
-            </Container>
+      <Container align="center" className={style.title}>
+        <Text
+          fontSize={20}
+          color="white"
+          fontWeight="semiBold"
+          lineHeight={1}
+          as="h2"
+        >
+          CONFIGURAÇÕES DE CONTA
+        </Text>
+      </Container>
 
-            {renderLinks()}
-          </Container>
-        </Hide>
+      <Container className={style.content}>
         <Container column gap={24}>
           <ChangeEmailCard />
-          <InactivateAccountCard />
-          <DeleteAccountCard />
+          <Container style={{ alignItems: 'initial' }} gap={24}>
+            <InactivateAccountCard />
+            <DeleteAccountCard />
+          </Container>
         </Container>
       </Container>
     </Container>